tests: Extract utility clients for K8s API
These were only defined in `test_volume`, which makes it harder for
other modules to re-use (even though they are quite handy!).
